Remove use of keyple from within control/domain
